Move File class to its own module
